requirements.txt
fastapi
uvicorn
transformers
torch
python-docx
PyMuPDF
python-multipart
